Move DatasourceConfig to api
