+++ impl/src/test/scala/quasar/impl/datasources/MockDatasourceConfigs.scala
@@ -18,7 +18,7 @@ package quasar.impl.datasources
 
 import slamdata.Predef.{Boolean, None, Option, Some, Unit}
 import quasar.api.ResourceName
-import quasar.api.datasource.DatasourceType
+import quasar.api.datasource.{DatasourceConfig, DatasourceType}
 import quasar.contrib.scalaz.MonadState_
 import quasar.fp.ski.κ2
 import MockDatasourceConfigs.{Configs, MonadConfigs}
